### 2023-12-12 #6142 @arfon
@Freakwill – you mentioned an earlier paper in your submission notes. Could you please link to it here so I can better understand what the focus/content of the earlier paper was?
