+++ examples/example.py
@@ -7,7 +7,7 @@ An ordinary example of the usage of `pyrimidine`
 from pyrimidine import MonoIndividual, BinaryChromosome, StandardPopulation
 from pyrimidine.benchmarks.optimization import *
 
-n_bags = 10
+n_bags = 50
 _evaluate = Knapsack.random(n_bags)  # : 0-1 array -> float
 
 # Define the individual class
@@ -25,10 +25,10 @@ class MyIndividual(MonoIndividual):
 # Define the population class
 class MyPopulation(StandardPopulation):
     element_class = MyIndividual
-    default_size = 6
+    default_size = 20
 
 """ Equiv. to
-    MyPopulation = StandardPopulation[MyIndividual] // 8
+    MyPopulation = StandardPopulation[MyIndividual] // 20
     or, as a population of chromosomes
     MyPopulation = StandardPopulation[(BinaryChromosome // n_bags).set_fitness(_evaluate)] // 8
 """
@@ -53,11 +53,11 @@ if __name__ == '__main__':
         'Mean Fitness': 'mean_fitness',
         'Best Fitness': 'max_fitness',
         'Standard Deviation of Fitnesses': 'std_fitness',
-        'number': lambda pop: len(pop.individuals)  # or `'n_individuals'`
+        # 'number': lambda pop: len(pop.individuals)  # or `'n_individuals'`
         }
 
     # Do statistical task and print the results through the evoluation
-    data = pop.evolve(stat=stat, n_iter=30, history=True, verbose=True)
+    data = pop.evolve(stat=stat, n_iter=100, history=True)
 
     # Visualize the data
     import matplotlib.pyplot as plt
@@ -66,8 +66,8 @@ if __name__ == '__main__':
     ax2 = ax.twinx()
     data[['Mean Fitness', 'Best Fitness']].plot(ax=ax)
     ax.legend(loc='upper left')
-    data['Standard Deviation of Fitnesses'].plot(ax=ax2, style='y-.')
-    ax2.legend(loc='lower right')
+    # data['Standard Deviation of Fitnesses'].plot(ax=ax2, style='y-.')
+    # ax2.legend(loc='lower right')
     ax.set_xlabel('Generations')
     ax.set_ylabel('Fitness')
     ax.set_title('Demo of solving the knapsack problem by GA')
