### 2023-12-13 #6142 @Freakwill
Hi @arfon
This is the paper link and code:
link: https://pan.baidu.com/s/1FKJ2TGtNLxQSoK3tESRP7g?pwd=t26e
This is the journal link: http://www.comprg.com.cn/post_show.asp?id=12947
It is in Chinese and the journal is not very high-end

> @Freakwill – you mentioned an earlier paper in your submission notes. Could you please link to it here so I can better understand what the focus/content of the earlier paper was?
