+++ examples/example.py
@@ -66,9 +66,9 @@ if __name__ == '__main__':
     ax2 = ax.twinx()
     data[['Mean Fitness', 'Best Fitness']].plot(ax=ax)
     ax.legend(loc='upper left')
-    # data['Standard Deviation of Fitnesses'].plot(ax=ax2, style='y-.')
-    # ax2.legend(loc='lower right')
+    data['Standard Deviation of Fitnesses'].plot(ax=ax2, style='y-.')
+    ax2.legend(loc='lower right')
     ax.set_xlabel('Generations')
     ax.set_ylabel('Fitness')
-    ax.set_title('Demo of solving the knapsack problem by GA')
+    # ax.set_title('Demo of solving the knapsack problem by GA')
     plt.show()
