### 2023-12-13 #6142 @arfon
Hi @Freakwill – I'm afraid we don't have a way to deal with non-English based submissions. Could you please provide some kind of translated artifact here for us to review? Thank you.
