### 2023-12-14 #6142 @Freakwill
I translate the article to English. It was written about 3 years ago.
@arfon Thank you for reviewing my paper.

[text.md](https://github.com/openjournals/joss-reviews/files/13668000/text.md)
