+++ docs/source/Release History.md
@@ -1,0 +1,8 @@
+# Release History
+
+## Current version
+
+- Give an example for "hybrid population", composed of populations and individuals
+
+## before v1.5
+?
