+++ text.md
@@ -1,0 +1,171 @@
+Journal: SOFTWARE DEVELOPMENT & APPLICATION
+date: 2021.1
+---
+
+# `Pyrimidine`: Object-Oriented Genetic Algorithm Framework in Python
+
+
+**Abstract**: `Pyrimidine` is a versatile framework developed for implementing genetic algorithms. It can also implement any iterative model, such as simulated annealing or particle swarm optimization. Its design is based on object-oriented programming, leveraging the metaprogramming capabilities of Python to treat populations as containers of individuals, individuals as containers of chromosomes, and containers as metaclasses for constructing different structures of individual and population classes. Due to its highly object-oriented nature, it is easily extensible and adaptable.
+
+**Keywords**: `Pyrimidine` framework; Genetic algorithm; Python language; Object-oriented programming; Metaprogramming
+
+## Introduction
+
+Genetic algorithms are a universal optimization method that mimics natural selection to solve optimization problems. They are among the earliest developed intelligent algorithms [1-3], widely used in various fields, and have been modified and combined with new algorithms [4]. This paper does not delve into the principles of genetic algorithms; for more details, refer to the literature [5,6] and references therein.
+
+Currently, several libraries in different languages provide frameworks for implementing genetic algorithms. Python, in particular, offers multiple genetic algorithm libraries, including the well-known DEAP [7], gaft, tpot [8,9] for machine learning parameter optimization, and scikit-opt and gplearn as extensions of scikit-learn [10]. This paper introduces the design of the general algorithm framework `Pyrimidine`. It strictly adheres to the object-oriented programming paradigm and utilizes the metaprogramming features of Python more rigorously than other libraries.
+
+## Framework Design
+
+The framework design primarily consists of two parts: classes implementing the fundamental concepts of genetic algorithms, such as individuals and populations, and metaclasses used to construct these classes.
+
+In the concrete implementation using Python, containers are object lists (or other feasible iterators). For instance, a population is a list of individuals; an individual is a list of chromosomes; and a chromosome is an array of genes. The specific implementation of the gene array is crucial and can be done using the standard library `array` or the well-known third-party numerical computing library numpy. The latter is more convenient for various numerical computations but has slower crossover operations.
+
+## Container Metaclass
+
+This metaclass is inspired by functional programming languages like Haskell and type theory. There are two main types of containers: lists and tuples, where lists represent elements with the same type, and tuples have no such restrictions. Users are free to modify the metaclass, so detailed content is not extensively covered here.
+
+Among them, `BaseIndividual` is the base class for all individual classes. This notation is influenced by variable types, such as a list of strings written as `List[String]`.
+
+## Basic Classes
+
[…]
