+++ pyrimidine/fa.py
@@ -11,54 +11,19 @@ from .chromosome import FloatChromosome
 from .individual import PolyIndividual
 from .utils import gauss, random
 
-from .deco import basic_memory
+from .pso import BaseParticle
 
 
 @basic_memory
-class BaseFirefly(BaseIndividual):
-
-    element_class = FloatChromosome
-    default_size = 3
+class BaseFirefly(BaseParticle):
 
     params = {
         'gamma': 1,
         'alpha': 1
     }
 
-    @property
-    def position(self):
-        raise NotImplementedError
-
-    @position.setter
-    def position(self, x):
-        raise NotImplementedError
-
-    @property
-    def velocity(self):
-        raise NotImplementedError
-
-    @velocity.setter
-    def velocity(self, v):
-        raise NotImplementedError
-
-    @property
-    def best_position(self):
-        # alias for the position of memory
-        if 'solution' not in self.memory or self.memory['solution'] is None:
-            return self.position
-        return self.memory['solution']
-
     def update_vilocity(self, fame=None, *args, **kwargs):
         raise NotImplementedError
-
-    def move(self, velocity=None):
-        if velocity is None:
-            self.position += self.velocity
-        else:
-            self.position += velocity
-
-    def decode(self):
-        return self.best_position
 
     def random_move(self):
         raise NotImplementedError
