+++ pyrimidine/learn/base.py
@@ -1,9 +1,0 @@
-
-
-class BaseEstimator:
-    def fit(self, X, y):
-        pop = self.config(X,y)
-        pop.evolve(n_iter=200, verbose=True)
-        self.best = pop.best_individual
-        self.postprocess()
-        return self
