+++ paper/paper.bib
@@ -19,7 +19,8 @@
     author = "Randal S. Olson and Ryan J. Urbanowicz and Peter C. Andrews and Nicole A. Lavender and La Creis Kidd and Jason H. Moore",
     journal = "Journal of Machine Learning Research",
     year = 2016,
-    pages = "123-137"
+    pages = "123-137",
+    doi="10.1007/978-3-319-31204-0_9"
 }
 
 @artical{le,
