+++ pyrimidine/chromosome.py
@@ -82,9 +82,25 @@ class NumpyArrayChromosome(BaseChromosome, np.ndarray):
         return "|".join(map(str, self))
 
     def cross(self, other):
-        # note that when len(self) == 2  ==>  k==1
+        """Crossover operation for a single chromosome
+
+        Note that when len(self) == 2  ==>  k==1
+        
+        Args:
+            other (BaseChromosome): another chromosome
+        
+        Returns:
+            BaseChromosome: new chromosome, as the child of the two chromosomes
+        """
+
         k = randint(1, len(self)-1)
         return self.__class__(np.concatenate((self[:k], other[k:]), axis=0))
+
+    def cross2(self, other):
+        # return 2 children after crossover
+        k = randint(1, len(self)-1)
+        return (self.__class__(np.concatenate((self[:k], other[k:]), axis=0)),
+                self.__class__(np.concatenate((other[:k], self[k:]), axis=0)))
 
     def copy(self, type_=None, *args, **kwargs):
         if type_ is None:
@@ -94,9 +110,17 @@ class NumpyArrayChromosome(BaseChromosome, np.ndarray):
         return obj
 
     def clone(self):
+        """Alias of `copy`, but regarded as a genetic operation
+        
+        Returns:
+            BaseChromosome
+        """
+
         return self.__class__(np.copy(self))
 
+    @side_affect
     def mutate(self, indep_prob=0.1):
+        # mutation
         for i in range(len(self)):
             if random() < indep_prob:
                 self[i] = self.element_class.random()
@@ -108,6 +132,7 @@ class VectorChromosome(NumpyArrayChromosome):
 
 class MatrixChromosome(NumpyArrayChromosome):
     
+    @side_affect
     def mutate(self, indep_prob=0.1):
         r, c = self.shape
         for i in range(r):
@@ -134,6 +159,7 @@ class NaturalChromosome(VectorChromosome):
 
     element_class = NaturalGene
 
+    @side_affect
     def mutate(self, indep_prob=0.1):
         for i in range(len(self)):
             if random()< indep_prob:
@@ -158,6 +184,7 @@ class BinaryChromosome(NaturalChromosome):
     def __str__(self):
         return "".join(map(str, self))
 
+    @side_affect
     def mutate(self, indep_prob=0.5):
         for i in range(len(self)):
             if random() < indep_prob:
@@ -185,6 +212,7 @@ class PermutationChromosome(NaturalChromosome):
         r = choice(rotations(self, other))
         rotate(self, r)
 
+    @side_affect
     def mutate(self):
         i, j = randint2(0, self.default_size-1)
         self[[i,j]] = self[[j,i]]
@@ -210,6 +238,7 @@ class FloatChromosome(NumpyArrayChromosome):
     def __str__(self):
         return "|".join(format(c, '.4') for c in self)
 
+    @side_affect
     def mutate(self, indep_prob=0.1, mu=0, sigma=None):
         sigma = sigma or self.sigma
         for i in range(len(self)):
@@ -358,7 +387,7 @@ class QuantumChromosome(CircleChromosome):
         return self.measure_result
 
 
-# Implement Chromosome class by array.array
+# Implement Chromosome class by `array.array`
 
 import copy
 import array
@@ -395,6 +424,7 @@ class ArrayChromosome(BaseChromosome, array.array):
     def copy(self, type_=None):
         return copy.copy(self)
 
+    @side_affect
     def mutate(self, indep_prob=0.1):
         a = self.random()
         for k in range(len(self)):
