+++ docs/conf.py
@@ -22,7 +22,7 @@ copyright = '2023, William Song'
 author = 'William Song'
 
 # The full version, including alpha/beta/rc tags
-release = '1.0'
+release = '1.5'
 
 
 # -- General configuration ---------------------------------------------------
