+++ examples/example-hybridpopulation.py
@@ -3,7 +3,7 @@
 from random import random
 import numpy as np
 
-from pyrimidine import MultiPopulation, HOFPopulation, MonoIndividual, BinaryChromosome
+from pyrimidine import HybridPopulation, HOFPopulation, BinaryChromosome
 from pyrimidine.benchmarks.optimization import *
 
 
@@ -11,57 +11,20 @@ from pyrimidine.benchmarks.optimization import *
 n_bags = 100
 _evaluate = Knapsack.random(n_bags)
 
-class _Individual(MonoIndividual[BinaryChromosome // n_bags]):
+_Individual = (BinaryChromosome // n_bags).set_fitness(_evaluate)
 
-    def decode(self):
-        return self[0]
 
-    def _fitness(self):
-        return _evaluate(self.decode())
+_Population = HOFPopulation[_Individual] // 5
 
-    @property
+
+class _HybridPopulation(HybridPopulation[_Population, _Population, _Individual, _Individual]):
+
     def max_fitness(self):
-        return self.fitness
-
-    @property
-    def mean_fitness(self):
-        return self.fitness
+        # compute maximum fitness for statistics
+        return max(self.get_all_fitness())
 
 
-class _Population(HOFPopulation):
-    element_class = _Individual
-    default_size = 10
-
-
-class HybridPopulation(MultiPopulation):
-    element_class = (_Population //2, _Individual//2)
-
-    def migrate(self, migrate_prob=None, copy=True):
-        migrate_prob = migrate_prob or self.migrate_prob
-        for this, other in zip(self[:-1], self[1:]):
-            if random() < migrate_prob:
-                if isinstance(this, _Population):
-                    this_best = this.get_best_individual(copy=copy)
-                    if isinstance(other, _Population):
-                        other.append(this.get_best_individual(copy=copy))
-                        this.append(other.get_best_individual(copy=copy))
-                    else:
-                        this.append(other.copy())
-                else:
-                    this_best = this.copy()
-                    if isinstance(other, _Population):
-                        other.append(this.get_best_individual(copy=copy))
-
-    def transition(self, *args, **kwargs):
-        for p in self:
-            p.transition(*args, **kwargs)
-        self.migrate()
-
-    # def max_fitness(self):
-    #     return max(e.fitness for e in self)
-
-
-sp = HybridPopulation.random()
+sp = _HybridPopulation.random()
 data = sp.evolve(n_iter=100, history=True)
 
 
