+++ docs/conf.py
@@ -38,7 +38,8 @@ extensions = [
     'sphinx.ext.todo',
     'sphinx.ext.coverage',
     'sphinx.ext.mathjax',
-    'sphinx.ext.autosectionlabel'
+    'sphinx.ext.autosectionlabel',
+    'sphinx.ext.napoleon',
     ]
 
 # Add any paths that contain templates here, relative to this directory.
