+++ tests/test_learn.py
@@ -3,7 +3,7 @@
 
 import numpy as np
 
-from pyrimidine.learn.neural_network import GAANN
+from pyrimidine.learn.neural_network import GAMLPRegressor
 from pyrimidine.learn.linear_regression import GALinearRegression
 
 
@@ -11,8 +11,8 @@ def test_ann():
     X = np.array([[0,0], [0,1], [1,0], [1,1]])
     Y = np.array([[0,1], [1,0], [1,0], [0,1]])
 
-    model = GAANN()
-    pop = GAANN.config(X, Y)
+    model = GAMLPRegressor()
+    pop = GAMLPRegressor.config(X, Y)
     s0 = pop.fitness
     model.max_iter = 2
     model.fit(X, Y, pop)
