+++ docs/index.rst
@@ -1,5 +1,5 @@
 .. pyrimidine documentation master file, created by
-   sphinx-quickstart on Sun Sep 20 13:05:18 2020.
+   sphinx-quickstart on Dec 18 2023.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
