### 2023-12-18 3a08e3f
paper; doc
