+++ pyrimidine/learn/linear_regression.py
@@ -8,8 +8,10 @@ import numpy.linalg as LA
 from sklearn.linear_model import LinearRegression
 from ..learn import BaseEstimator
 
+from ..chromosome import BinaryChromosome
 from ..population import StandardPopulation
 from digit_converter import IntervalConverter
+
 
 c = IntervalConverter(lb=-60, ub=60)
 class _BinaryChromosome(BinaryChromosome):
