+++ tests/test_individual.py
@@ -5,15 +5,16 @@ import numpy as np
 from pyrimidine.individual import *
 
 
-class TestChromosome:
+class TestIndividual:
 
     def test_make(self):
         b = makeIndividual(element_class=BinaryChromosome, n_chromosomes=2, size=8).random()
         assert len(b) == 2
 
     def test_make_binary(self):
-        b = makeBinaryIndividual(size=(8, 4)).random()
-        assert len(b[0]) == 8 and len(b[1]) == 4
+        b = makeBinaryIndividual(size=(6, 4)).random()
+        print(b)
+        assert len(b[0]) == 6 and len(b[1]) == 4
 
     def test_cross(self):
         I = MonoIndividual[BinaryChromosome]
