+++ paper/paper.bib
@@ -174,12 +174,25 @@ pages="560--564"
 }
 
 @software{neat-python,
-author = {McIntyre, Alan and Kallada, Matt and Miguel, Cesar G. and Feher de Silva, Carolina and Netto, Marcio Lobo},
-title = {{neat-python}}
+    author = {McIntyre, Alan and Kallada, Matt and Miguel, Cesar G. and Feher de Silva, Carolina and Netto, Marcio Lobo},
+    title = {neat-python}
 }
 
 @online{algebra,
-title = "Algebraic structure",
-url = "https://en.wikipedia.org/wiki/Algebraic_structure",
-year = 2023
+    title = "Algebraic structure",
+    url = "https://en.wikipedia.org/wiki/Algebraic_structure",
+    year = 2023
+}
+
+@article{radtke,
+    doi = {10.21105/joss.01701},
+    url = {https://doi.org/10.21105/joss.01701},
+    year = {2020},
+    publisher = {The Open Journal},
+    volume = {5},
+    number = {47},
+    pages = {1701},
+    author = {Jonas Joacir Radtke and Guilherme Bertoldo and Carlos Henrique Marchi},
+    title = {DEPP - Differential Evolution Parallel Program},
+    journal = {Journal of Open Source Software}
 }
