### 2023-12-20 0122bdc
paper
