### 2023-12-21 #6142 @arfon
> I translate the article to English. It was written about 3 years ago. In these three years, the program has been greatly redesigned.

Many thanks. Refactors/redesign of software doesn't necessarily mean a follow-on publication in JOSS would be permitted by us. Could you say more about what significant new capabilities have been added since the original release/publication, and what sorts of research challenges can be addressed by them?

I'm also curious to learn more about community use here. Does the original paper have many citations? Are people using this package for their research already?
