+++ docs/conf.py
@@ -12,7 +12,7 @@
 #
 import os
 import sys
-sys.path.insert(0, os.path.abspath('.../pyrimidine'))
+sys.path.insert(0, '..')
 
 
 # -- Project information -----------------------------------------------------
