+++ pyrimidine/benchmarks/neural_network.py
@@ -5,7 +5,7 @@ import numpy.linalg as LA
 
 from ..benchmarks import BaseProblem
 
-from .utils import max0
+from ..utils import max0
 
 
 relu = max0
