+++ docs/build/_sources/source/Release History.md.txt
@@ -4,6 +4,8 @@
 
 - Give an example for "hybrid population", composed of populations and individuals
 - parallel computing, limited to computing the fitnesses parallely
+- add class method `solve`, use `Population.solve` to get the solution in a convenient way, where `Population` is the class for any iterative algorithm.
+- update the decorator for cache
 
 ## before v1.5
 ?

[…]
