+++ pyrimidine/optimize.py
@@ -8,7 +8,7 @@ import numpy as np
 
 from .chromosome import BinaryChromosome
 from .individual import makeIndividual, makeBinaryIndividual
-from .population import HOFPopulation
+from .population import StandardPopulation
 from .de import DifferentialEvolution
 
 
@@ -36,41 +36,61 @@ def _decode(c, a, b):
     return IntervalConverter(a, b)(c)
 
 
-def ga_minimize(func, *xlim, decode=_decode, population_size=20, size=8):
+def _make_individual(*xlim, size=8):
+    """Make an individual class
+    
+    Args:
+        *xlim: intervals
+        size (int, tuple of int): the length of the encoding of xi
+    """
+
+    ndim = len(xlim)
+
+    if ndim == 1:
+        class _Individual(BinaryChromosome // size):
+
+            def _fitness(self):
+                return - func(self.decode())
+
+            def decode(self):
+                return decode(self, *xlim)
+    else:
+        class _Individual(makeIndividual(n_chromosomes=ndim, size=size)):
+            default_size = ndim
+
+            def _fitness(self):
+                return - func(self.decode())
+
+            def decode(self):
+                return np.asarray([decode(c, a, b) for c, (a, b) in zip(self, xlim)])
+
+    return _Individual
+
+
+def ga_minimize(func, *xlim, decode=_decode, population_size=20, size=8, *args, **kwargs):
     """
     GA(with hall of fame) for minimizing the function `func` defined on `xlim`
 
     Arguments:
         func: objective function defined on R^n
-        xlim: the intervals of xi
-        decode: transform a binary sequence to a real number
+        xlim {tuple of number pairs}: the intervals of xi
+        decode {mapping}: transform a binary sequence to a real number
             ('0-1' sequence, lower_bound, upper_bound) |-> xi
-        population_size: size of population
-        size: the length of the encoding of xi
+        population_size {int}: size of the population
+        size {int or tuple of int}: the length of the encoding of xi
 
     Example:
         ga_minimize(lambda x:x[0]**2+x[1], (-1,1), (-1,1))
     """
 
-    class _Individual(makeIndividual(n_chromosomes=len(xlim), size=size)):
-
-        def _fitness(self):
-            return - func(self.decode())
-
-        def decode(self):
-            return np.asarray([decode(c, a, b) for c, (a, b) in zip(self, xlim)])
-
-    class _Population(HOFPopulation):
-        element_class = _Individual
-        default_size = population_size
+    _Individual = _make_individual(*xlim, size)
+    _Population = StandardPopulation[_Individual] // population_size
 
     pop = _Population.random()
-    pop.ezolve()
-
-    return pop.solution
+    return pop.ezolve(*args, **kwargs).solution
 
 
-def de_minimize(func, *xlim, decode=_decode, population_size=20, size=8):
+def de_minimize(func, *xlim, decode=_decode, population_size=20, size=8, **kwargs):
     """
     DE for minimizing the function `func` defined on `xlim`
 
@@ -79,73 +99,63 @@ def de_minimize(func, *xlim, decode=_decode, population_size=20, size=8):
         xlim: the intervals of xi
         decode: transform a binary sequence to a real number
             ('0-1' sequence, lower_bound, upper_bound) |-> xi
-        population_size: size of population
+        population_size: size of the population
         size: the length of the encoding of xi
 
     Example:
         ga_minimize(lambda x:x[0]**2+x[1], (-1,1), (-1,1))
     """
 
-    class _Individual(makeIndividual(n_chromosomes=len(xlim), size=size)):
-        default_size = len(xlim)
-
-        def _fitness(self):
-            return - func(self.decode())
-
-        def decode(self):
[…]
