### 2024-01-05 #6142 @Freakwill
> > I translate the article to English. It was written about 3 years ago. In these three years, the program has been greatly redesigned.
>
> Many thanks. Refactors/redesign of software doesn't necessarily mean a follow-on publication in JOSS would be permitted by us. Could you say more about what significant new capabilities have been added since the original release/publication, and what sorts of research challenges can be addressed by them?
>
> I'm also curious to learn more about community use here. Does the original paper have many citations? Are people using this package for their research already?

In addition to refactoring the original code, we have adopted algebraic programming more thoroughly, greatly improving scalability, and making it easy to any iterative simulation. It has a natural way to implement the multi-population genetic algorithms (even multi-multi-population), since we make use of the similar algebraic structures of the two. We have implemented a variety of intelligent algorithms, including quantum GA, PSO, DE; designed specialized chromosomes for special variables such as probability distributions and permutations; designed a caching system to accelerate computation; and initially implemented parallel computing, although currently limited to parallel computation of fitness. Combined with machine learning models, we provide scikit-learn-style APIs, such as the `GAMLPRegressor` with the same APIs of `MLPRegressor`, which implements the evoluationary neural network.

Currently, this package has not been officially applied to research. However, I have used it to solve real-world problems (such as [movie scheduling](https://github.com/Freakwill/movie-schedule)). We have produced instructional [videos](https://www.youtube.com/watch?v=uVf3y427ei4&t=3s) to allow more people to understand this package. I am also persuading my colleagues to use this package.

The journal I published had little influence in itself; my papers also seemed to be cited by no one. That was my first attempt to let people know about my immature works.
