### 2024-01-09 #6142 @arfon
> In addition to refactoring the original code, we have adopted algebraic programming more thoroughly, greatly improving scalability, and making it easy to any iterative simulation. It has a natural way to implement the multi-population genetic algorithms (even multi-multi-population), since we make use of the similar algebraic structures of the two. We have implemented a variety of intelligent algorithms, including quantum GA, PSO, DE; designed specialized chromosomes for special variables such as probability distributions and permutations; designed a caching system to accelerate computation; and initially implemented parallel computing, although currently limited to parallel computation of fitness. Combined with machine learning models, we provide scikit-learn-style APIs, such as the GAMLPRegressor with the same APIs of MLPRegressor, which implements the evoluationary neural network.

OK, thanks for the additional information @Freakwill. Have all of these changes been made since the earlier paper was published?
