### 2024-03-04 #6142 @boisgera
> @editorialbot invite @boisgera as editor
>
> 👋 @boisgera – would you be willing to edit this submission for JOSS?

Yes, sure!
