### 2024-04-02 #6575 @editorialbot
[REVIEW]: Pyrimidine: An algebra-inspired Programming framework for evolutionary algorithms
**Submitting author:** @Freakwill (<a href="http://orcid.org/0000-0002-4409-7276">Congwei Song</a>)
**Repository:** https://github.com/Freakwill/pyrimidine
**Branch with paper.md** (empty if default branch):
**Version:** 1.8
