### 2024-04-22 #6575 @mmore500
Some comments on the manuscript.

## Title block
Remove “These authors contributed equally” for one author.

## Summary
add a space in “programming(OOP)”

## Statement of need
- citations to holland and katoch are missing
- add a space in “genetic algorithm(GA)”
- be perhaps “optimization” or “sophisticated” would be more appropriate than “intelligent” algorithms”?
- Consider overviewing features/modules in the library around line 37 (end of statement of need )—— in addition to optimization algorithms: what’s in benchmarks, what chromosome models are implemented, what crossover models are implemented, etc.

## Algebra-inspired programming
- The advantage or purpose of the algebra-inspired approach is never directly stated. This should be explained more explicitly.
- “In a typical Python implementation, …” [implementation of what?]
- Prefer citations over hyperlinks (i.e., for NumPy)
- The meaning of resp. is unclear
- “discrepancy” should probably be “difference”
- Around line 52, the mixing of Python syntax `S[ ]` with formal mathematical notation is a bit confusing.
- How line 98, 99, and the following code segment relate to the preceding manuscript narrative is unclear (beginning with “Algebraically, there is no discrepancy between…”)

## Figure 1:
 — the twin y axis on right isn’t labeled
 — caption does not need to mention that plotting algorithms are not included in the library

## Visualization:
Because visualization isn’t a core feature of the library, consider rewriting the visualization section to exclude the plotting and focus on history extraction. The majority of the example code in this section is generic boilerplate matplotlib.

## An example to begin
- The variables c, x w and W appearing in the display math are not defined in the text

## Conclusion
In closing sentence,
- GitHub doesn’t need to be linked
- might mention software license
- numerous examples [showing what?]

## References
- avoid citations to wikipedia
- type hints do not need to be cited

## Overall
- The purpose of overriding `_fitness` should be mentioned in the accompanying text. The override of the // operator to denote size should also similarly explicitly explained.
- there are several broken references (search the manuscript for “??”)
