### 2024-05-03 #6575 @boisgera
Hi @Freakwill,

@mmore500 has made some comments/requests about your paper ~2 weeks ago (see https://github.com/openjournals/joss-reviews/issues/6575#issuecomment-2071073010)

Would it help if I transcribed these bullet points as checklists and register them as issues in your project repository?
