### 2024-05-08 c780846
Clean up some typos, styling in README
