+++ README.md
@@ -1,6 +1,6 @@
 # pyrimidine
 
-`pyrimidine` is an extensible framework of genetic/evolutionary algorithm by Python. See [pyrimidine's document](https://pyrimidine.readthedocs.io/) for more details.
+`pyrimidine` is an extensible framework of genetic/evolutionary algorithm by Python. See [pyrimidine's documentation](https://pyrimidine.readthedocs.io/) for more details.
 
 ![LOGO](logo.png)
 
@@ -8,7 +8,7 @@
 
 > -- Why is the package named as “pyrimidine”? 
 > -- Because it begins with “py”.
-> -- Are you kiding?
+> -- Are you kidding?
 > -- No, I am serious.
 
 
@@ -20,7 +20,7 @@ It has been uploaded to [pypi](https://pypi.org/project/pyrimidine/), so downloa
 
 ## Video tutorials
 
-[An gental introduction](https://www.youtube.com/watch?v=uVf3y427ei4&t=3s)
+[An gentle introduction](https://www.youtube.com/watch?v=uVf3y427ei4&t=3s)
 
 ## Idea
 
@@ -38,7 +38,7 @@ class BasePopulation(PopulationModel, metaclass=MetaContainer):
     default_size = 20
 ```
 
-There is two main kinds of containers: list-like and tuple-like. See following examples.
+There are two main kinds of containers: list-like and tuple-like. See the following examples.
 
 ```python
 # individual with chromosomes of type _Chromosome
@@ -57,7 +57,7 @@ s={a:A}:S
 
 We could define a population as a container of individuals or chromosomes, and an individual is a container of chromosomes.
 
-Algebraically, an indivdiual has only one chromosome is equivalent to a chromosome mathematically. A population could also be a container of chromosomes. If the individual has only one chromosome, then just build the population based on chromosomes directly.
+Algebraically, an individual with one chromosome is equivalent to a chromosome mathematically. A population could also be a container of chromosomes. If the individual has only one chromosome, then just build the population based on chromosomes directly.
 
 The methods are the functions or operators defined on $s$.
 
@@ -69,12 +69,12 @@ The methods are the functions or operators defined on $s$.
 - BaseChromosome: sequence of genes, represents part of a solution
 - BaseIndividual: sequence of chromosomes, represents a solution of a problem
 - BasePopulation: a container of individuals, represents a container of a problem
-                also the state of a stachostic process
-- BaseMultipopulation: a container of population for more complicated optimalization
+                also the state of a stochastic process
+- BaseMultipopulation: a container of population for more complicated optimization
 
 
 ### import
-Just use the command `from pyrimidine import *` import all of the algorithms.
+Just use the command `from pyrimidine import *` to import all of the algorithms.
 
 ### subclass
 
@@ -90,7 +90,7 @@ just subclass `MonoIndividual` in most cases.
 ```python
 class MyIndividual(MonoIndividual):
     """individual with only one chromosome
-    we set the gene is 0 or 1 in the chromosome
+    we set the gene to 0 or 1 in the chromosome
     """
     element_class = BinaryChromosome
 
@@ -98,11 +98,11 @@ class MyIndividual(MonoIndividual):
         ...
 ```
 
-Since the helper `makeIndividual(n_chromosomes=1, size=8)` could create such individual, it is equivalent to
+Since the helper `makeIndividual(n_chromosomes=1, size=8)` could create such an individual, it is equivalent to
 
 ```python
 class MyIndividual(binaryIndividual()):
-    # only need define the fitness
+    # only need to define the fitness
     def _fitness(self):
         ...
 ```
@@ -118,8 +118,8 @@ class _Chromosome(BinaryChromosome):
     def decode(self):
         """Decode a binary chromosome
         
[…]
