### 2024-05-08 Freakwill/pyrimidine#4 @mmore500
pip install does not install required dependencies
```bash
python3 -m pip install pyrimidine
python3 -c "import pyrimidine;"
```

gives the error message

```
Traceback (most recent call last):
  File "<string>", line 1, in <module>
  File "/home/mmore500/.local/lib/python3.12/site-packages/pyrimidine/__init__.py", line 4, in <module>
    from .base import *
  File "/home/mmore500/.local/lib/python3.12/site-packages/pyrimidine/base.py", line 62, in <module>
    from toolz import concat
ModuleNotFoundError: No module named 'toolz
```

Ideally, required dependencies would be listed in `pyproject.toml` in order to ensure that they are installed alongside pyrimidine so it is usable without further manual installations.
