### 2024-05-08 Freakwill/pyrimidine#4 @mmore500
This issue is part of JOSS review https://github.com/openjournals/joss-reviews/issues/6575
