+++ examples/example2.py
@@ -69,14 +69,14 @@ class MyPopulation(HOFPopulation[MyIndividual]):
 if __name__ == '__main__':
 
     stat = {'Mean Fitness':'mean_fitness',
-        'Best Fitness': 'best_fitness'}
+        'Best Fitness': 'max_fitness'}
 
     import matplotlib.pyplot as plt
     fig = plt.figure()
     ax = fig.add_subplot(111)
     pop = MyPopulation.random(n_individuals=40, size=[8]*ndim+[8])
-    cpy = pop.clone(HOFPopulation[ExampleIndividual])
-    d = cpy.evolve(n_iter=200, stat=stat, history=True)
+    cpy = pop.copy(HOFPopulation[ExampleIndividual])
+    d = cpy.evolve(n_iter=200, stat=stat, history=True, verbose=True)
     ax.plot(d.index, d['Mean Fitness'], d.index, d['Best Fitness'], '.-')
     d = pop.evolve(n_iter=200, stat=stat, history=True)
     ax.plot(d.index, d['Mean Fitness'], d.index, d['Best Fitness'], '.-')
