+++ docs/build/html/_sources/Misc.md.txt
@@ -1,0 +1,28 @@
+# Misc
+
+## Pictures
+Generated Picture
+![](taichi.png)
+
+Original Picture
+![](https://img.sj33.cn/uploads/allimg/201404/7-1404160TQ63V.png)
+
+## Videos
+<video
+src="tsp.mp4" controls=""
+height=350 
+width=600> 
+TSP solved by GA
+</video>
+
+<video
+src="https://youtu.be/KwpNPCNutcI" controls=""
+height=350 
+width=600> 
+</video>
+
+If it fails to play, then access the url https://youtu.be/KwpNPCNutcI
+
+## Tutorial
+
+https://youtu.be/LWc2V28kzgA

[…]
