+++ docs/source/conf.py
@@ -39,7 +39,7 @@ extensions = [
     'sphinx.ext.coverage',
     'sphinx.ext.mathjax',
     'sphinx.ext.autosectionlabel',
-    'sphinx.ext.napoleon',
+    'sphinx.ext.napoleon'
     ]
 
 # Add any paths that contain templates here, relative to this directory.
