+++ pyrimidine/ep.py
@@ -6,18 +6,20 @@
 Invented by L. Fogel[1966] for designing FSM initially.
 
 General form of the algorithm:
-initialize a population with N individuals
-loop:
-    calculate f(x) for each x in population
-    mutate x for each x
-    get new population (mixed with the original population)
-    select best N individuals from the 2N mixed population
 
-The mutation:
-    x' <- x + r*sqrt(v)
-    v' <- v + c*r*sqrt(v) (make sure that v>epsilon)
+    initialize a population with N individuals
+    loop:
+        calculate f(x) for each x in population
+        mutate x for each x
+        get new population (mixed with the original population)
+        select best N individuals from the 2N mixed population
 
-Caution: No cross operation in EP
+    The mutation:
+        x' <- x + r*sqrt(v)
+        v' <- v + c*r*sqrt(v) (make sure that v>epsilon)
+
+Remark:
+    No cross operation in EP
 """
 
 from random import choice
