### 2024-05-12 #6575 @boisgera
For the record, I have e-mailed the project author to make sure that the communication channels (this JOSS issue as well as the project issue) with him are open and asked him to try to publicly acknowledge the reviewer messages (even if the issue can't be adressed right away).
