+++ README.md
@@ -11,7 +11,6 @@
 > -- Are you kiding?
 > -- No, I am serious.
 
-
 If you have more questions, then log in [google group](https://groups.google.com/g/pyrimidine) and post your questions.
 
 ## Download
@@ -74,9 +73,30 @@ The methods are the functions or operators defined on $s$.
 
 
 ### import
-Just use the command `from pyrimidine import *` import all of the algorithms.
 
-### subclass
+To import all algorithms for beginners, simply use the command `from pyrimidine import *`.
+
+To speed the lib, use the following commands 
+```
+from pyrimidine import BaseChromosome, BaseIndividual, BasePopulation # import the base classes form `base.py` to build your own classes
+
+# Commands used frequently
+from pyrimidine.base import BinaryChromosome, FloatChromosome # import the Chromosome classes and utilize them directly
+# equivalent to `from pyrimidine import BinaryChromosome, FloatChromosome`
+from pyrimidine.population import StandardPopulation, HOFPopulation # For creating population with standard GA 
+# the same effect with `from pyrimidine import StandardPopulation, HOFPopulation`
+from pyrimidine.indiviual import makeIndividual # a helper to make Individual objects, or `from pyrimidine import makeIndividual`
+
+from pyrimidine import MultiPopulation # build the multi-populations
+from pyrimidine import MetaContainer # meta class for socalled container class, that is recommended to be used for creating novel evolutionary algorithms.
+
+from pyrimidine.deco import * # import all decorators
+from pyrimidine.deco import fitness_cache, basic_memory # use the cache decorator and memory decorator
+
+from pyrimidine import optimize # do optimization implictly with GAs
+```
+
+### subclasses
 
 #### Chromosome
 
@@ -88,6 +108,11 @@ As an array of 0-1s, `BinaryChromosome` is used most frequently.
 just subclass `MonoIndividual` in most cases.
 
 ```python
+from pyrimidine.individual import MonoIndividual
+from pyrimidine.chromosome import BinaryChromosome
+
+# or from pyrimidine import MonoIndividual, BinaryChromosome
+
 class MyIndividual(MonoIndividual):
     """individual with only one chromosome
     we set the gene is 0 or 1 in the chromosome
@@ -101,19 +126,22 @@ class MyIndividual(MonoIndividual):
 Since the helper `makeIndividual(n_chromosomes=1, size=8)` could create such individual, it is equivalent to
 
 ```python
+from pyrimidine.individual import binaryIndividual
+
 class MyIndividual(binaryIndividual()):
     # only need define the fitness
     def _fitness(self):
         ...
 ```
 
-
 If an individual contains several chromosomes, then subclass `MultiIndividual` or `PolyIndividual`. It could be applied in multi-real-variable optimization problems where each variable has a separative binary encoding.
-
 
 In most cases, we have to decode chromosomes to real numbers.
 
 ```python
+from pyrimidine.individual import BaseIndividual
+from pyrimidine.chromosome import BinaryChromosome
+
 class _Chromosome(BinaryChromosome):
     def decode(self):
         """Decode a binary chromosome
@@ -134,6 +162,8 @@ class ExampleIndividual(BaseIndividual):
 If the chromosomes in an individual are different with each other, then subclass `MixedIndividual`, meanwhile, the property `element_class` should be assigned with a tuple of classes for each chromosome.
 
 ```python
+from pyrimidine.individual import MixedIndividual
+
 class MyIndividual(MixedIndividual):
     """
     Inherit the fitness from ExampleIndividual directly.
[…]
