+++ pyrimidine/__init__.py
@@ -18,7 +18,7 @@ from .de import *
 from .ba import *
 
 
-__version__ = "1.6"
+__version__ = "1.6.1"
 
 __template__ = """
 from pyrimidine.chromosome import BinaryChromosome
