+++ paper/paper.bib
@@ -178,12 +178,6 @@ pages="560--564"
     title = {neat-python}
 }
 
-@online{algebra,
-    title = "Algebraic structure",
-    url = "https://en.wikipedia.org/wiki/Algebraic_structure",
-    year = 2023
-}
-
 @article{radtke,
     doi = {10.21105/joss.01701},
     url = {https://doi.org/10.21105/joss.01701},
