+++ paper/paper.bib
@@ -41,6 +41,7 @@
 }
 
 @online{numpy,
+    author = {NumPy Developers},
     title="Numpy - The fundamental package for scientific computing with Python",
     url="https://numpy.org/",
     publisher="NumPy Developers",
