+++ paper/paper.bib
@@ -41,7 +41,7 @@
 }
 
 @online{numpy,
-    author = {NumPy Developers},
+    author = {{NumPy Developers}},
     title="Numpy - The fundamental package for scientific computing with Python",
     url="https://numpy.org/",
     publisher="NumPy Developers",
@@ -97,7 +97,6 @@
     eprint = {https://academic.oup.com/gji/article-pdf/236/1/537/53710009/ggad446.pdf},
 }
 
-
 @article{supasil,
 doi = {10.1088/1742-6596/1719/1/012102},
 url = {https://dx.doi.org/10.1088/1742-6596/1719/1/012102},
@@ -146,7 +145,6 @@ pages="560--564"
       archivePrefix={arXiv},
       primaryClass={astro-ph.IM}
 }
-
 
 @article{wang,
   title = {Particle Swarm Optimization Algorithm: An Overview},
