+++ paper/paper.md
@@ -85,10 +85,10 @@ We introduce the concept of a **container**, simulating an **(algebraic) system*
 
 A container $s$ of type $S$, with elements of type $A$, is represented by following expression:
 
-$$
-s = \{a:A\}: S \quad \text{or} \quad s:S[A] 
-\label{container}
-$$
+    $$
+    s = \{a:A\}: S \quad \text{or} \quad s:S[A] 
+    \label{container}
+    $$
 
 where the symbol $\{\cdot\}$ signifies either a set, or a sequence to emphasize the order of the elements. The notation $S[A]$ mimicks Python syntax, borrowed from the module [typing](https://docs.python.org/3.11/library/typing.html?highlight=typing#module-typing).
 
