### 2024-06-06 f462559
n_iter --> max_iter
