+++ examples/comparison-proba.py
@@ -59,8 +59,8 @@ class YourPopulation(_Mixin, HOFPopulation):
 mypop = MyPopulation.random()
 for i in mypop: i.measure()
 yourpop = MyPopulation([YourIndividual(i.measure_result) for i in mypop])
-mydata = mypop.evolve(n_iter=50, history=True)
-yourdata = yourpop.evolve(n_iter=50, history=True)
+mydata = mypop.evolve(max_iter=50, history=True)
+yourdata = yourpop.evolve(max_iter=50, history=True)
 
 import matplotlib.pyplot as plt
 fig = plt.figure()
