### 2024-06-09 #6575 @boisgera
> I use the following cross-references according to [the docs of joss](https://joss.readthedocs.io/en/latest/paper.html#tables-and-figures). But why dose it result in `??`?
>
> ```
> A concise comparison between `pyrimidine` and several popular frameworks provided in \autoref{frameworks}, such as ....
>
> : Comparison of the popular genetic algorithm frameworks. []{label="frameworks"}
> ```

I see that you apparently still have some weird formatting in this line, I'll have a look.
