### 2024-06-09 #6575 @boisgera
AFAICT the `autoref{eq:container}` issue should be solved if you use the syntax

```
\begin{equation} \label{eq:container}
s = \{a:A\}: S \quad \text{or} \quad s:S[A]
\end{equation}
```
