### 2024-06-09 #6575 @boisgera
`autoref{history}` already works as expected :+1:
