### 2024-06-09 #6575 @boisgera
And with respect to `autoref{frameworks}`, the caption

```
: Comparison of the popular genetic algorithm frameworks. \label{frameworks}
```

shoud appear **after** the table, not before. These are the three use of `autoref` that I can find in the paper.
