+++ pyrimidine/mixin.py
@@ -18,7 +18,8 @@ IterativeMixin  --->  CollectiveMixin
 FitnessMixin  --->  PopulationMixin
 """
 
-
+from deprecated import deprecated
+from warnings import deprecated
 from operator import methodcaller, attrgetter
 
 import numpy as np
@@ -37,7 +38,7 @@ from .errors import *
 class IterativeMixin:
     # Mixin class for iterative algrithms
 
-    params = {'max_iter': 100}
+    params = {'n_iter': 100}
 
     # @property
     # def _row(self):
@@ -62,19 +63,19 @@ class IterativeMixin:
         """
         raise NotImplementedError('If you apply a local search algorithm, you must define the `local_search` method.')
 
-    def ezolve(self, max_iter=None, init=True):
+    def ezolve(self, n_iter=None, init=True):
         # Extreamly eazy evolution method for lazybones
-        max_iter = max_iter or self.max_iter
+        n_iter = n_iter or self.n_iter
         if init:
             self.init()
-        for k in range(1, max_iter+1):
+        for k in range(1, n_iter+1):
             self.transition(k)
 
-    def evolve(self, initialize:bool=True, max_iter:int=100, period:int=1, verbose:bool=False, history=False, stat=None, attrs=('solution',), control=None):
+    def evolve(self, initialize:bool=True, n_iter:int=100, period:int=1, verbose:bool=False, history=False, stat=None, attrs=('solution',), control=None):
         """Get the history of the whole evolution
 
         Keyword Arguments:
-            max_iter {number} -- number of iterations (default: {None})
+            n_iter {number} -- number of iterations (default: {None})
             period {integer} -- the peroid of stat
             verbose {bool} -- to print the iteration process
             stat {dict} -- a dict(key: function mapping from the object to a number) of statistics 
@@ -88,7 +89,7 @@ class IterativeMixin:
 
         assert control is None or callable(control)
  
-        max_iter = max_iter or self.max_iter
+        n_iter = n_iter or self.n_iter
 
         if isinstance(stat, dict): stat = Statistics(stat)
         
@@ -117,7 +118,7 @@ class IterativeMixin:
 -------------------------------------------------------------""")
             print(_row(0, attrs, res))
 
-        for t in range(1, max_iter+1):
+        for t in range(1, n_iter+1):
             self.transition(t)
 
             if history_flag and (period == 1 or t % period ==0):
@@ -174,6 +175,7 @@ class IterativeMixin:
         return self.copy()
 
     def decode(self):
+        # decode the object to the real solution
         return self
 
     @classmethod
@@ -351,17 +353,19 @@ class CollectiveMixin(IterativeMixin):
             return self[ks]
 
     def observe(self, name='_system'):
-        for o in self:
-            setattr(o, name, self)
-
-    @property
-    def op(self):
-        class _C:
-            def __getitem__(obj, s):
-                def _f(*args, **kwargs):
-                    return getattr(self._system, s)(self, *args, **kwargs)
-                return _f
-        return _C()
+        """Let the elements observe the system, by `o._system`
+        
+        Args:
+            name (str, optional): the attribute name of the system
+        """
+        for e in self:
+            setattr(e, name, self)
+            class _C:
+                def __getitem__(obj, s):
+                    def _f(*args, **kwargs):
+                        return getattr(self._system, s)(self, *args, **kwargs)
+                    return _f
+            e.op = _C()
 
 
 class PopulationMixin(FitnessMixin, CollectiveMixin):
@@ -414,10 +418,10 @@ class PopulationMixin(FitnessMixin, CollectiveMixin):
     @property
     def std_fitness(self):
         return np.std(self.get_all_fitness())
-
+    
     @property
[…]

[…]
