### 2024-07-07 #6575 @boisgera
Hi @Freakwill,

Nice to see that you have solved the reference issues in the article!

Could you comment/move forward the 3 open issues/PR that are still flagged as open so that the reviewers (@mmore500, @sjvrijn ) can close them move the review forward?  For example, I have tested `pip install` again (https://github.com/Freakwill/pyrimidine/issues/4) and (at least in my configuration), you have solved this issue. Could you update the issue to make that progress explicit?

Kind regards,

Sébastien
