### 2024-07-22 #6575 @boisgera
Hi @mmore500, @sjvrijn,

I see that the previous issues/PR have been handled :+1:

Can you now address the **Functionality** section of your checklist?

Cheers,

Sébastien
