### 2024-08-05 #6575 @mmore500
Thank you for your revisions to the manuscript --- it has improved greatly!! A few comments from my current read through.

----

I am unsure about the `formatting` of Pyrimidine in the title, as I don't see other JOSS papers with the title formatted this way.
I didn't find any guidance in this regard in the JOSS Paper format guidelines.
If this won't cause problems later, it's fine to keep it as is, but I just wanted to flag it so it could be fixed now if necessary (perhaps @boisgera could advise on this?).

---

On the topic of issues that may arise later on, some attention should be paid to the word count of the manuscript.
According to the JOSS Paper format guidelines,
> The paper should be between 250-1000 words. Authors submitting papers significantly longer than 1000 words may be asked to reduce the length of their paper.

The first draft of the present manuscript was around 1,800 words.
Currently, the length is around 2,000 words.
In terms of where to trim, I would focus on making content on pages 3, 4, and 5 more concise.

---

> ... stands as a versatile framework designed for GAs, offering exceptional extensibility...
GA should be defined when first used (it's currently defined below at the second usage)

Table 1: Some references are in main text, and others are in table. I would recommend moving all references out of table and into main text.

---

>  So far, We have implemented a variety of intelligent algorithms via...

- Fix capitilization on "We."
- Consider using a more precise term than "intelligent" --- could "optimization" be a better term here? Or "sophisticated"?
- A reference for evolutionary programming would be helpful. Perhaps also one for simulated annealing (although this is sufficiently common to not be strictly necessary).

---

> This library provides a wide range of chromosome classes in the chromosome module...

The term chromosome has not yet been defined.
I infer that this should be related to the genetic representation used by these algorithms, but a sentence more explicitly introducing this term and its meaning in this context would be helpful.

---

> It should not be confused with so-called algebraic programming, but it draws inspiration...

A citation with some information on algebraic programming would be helpful here.

---

> It has better scalability.

Is this meant in the sense of parallel/distributed computing?
Or in the sense of scaling being highly extensible?

---

> Each class implements corresponding genetic operations such as crossover and others.

It would be helpful to have the "others" listed out here (or at least two or three more listed if there are many).

---

> Represented by following expression:

should read as "Represented by the following expression:"

---

> The lifting of a function/method...

It seems to be implied that this lifting idea is related to the extensibility of Pyrimidine to multi-population schemes.
This could be clarified, particularly with respect to selection operations (i.e., I assume that subpopulations are not replicated/selected as individual organisms are?) and with respect to migration operations between populations (are these implemented as crossover operations?).

---

Where the display math (i.e., center-aligned) ends a sentence in line 86, it should be formatted with punctuation. See [here](https://tex.stackexchange.com/a/170691/316176).

---

The term "super-metaclass," while technically correct, may be a bit difficult to parse for readers without a background in type theory.

---

> There are three base classes in pyrimidine: `BaseChromosome`, `BaseIndividual`, `BasePopulation`, to create chromosomes, individuals and populations respectively.

Missing an "and."
Additionally, while the intended role of these base classes within the evolutionary algorithm it is sufficiently clear, the difference in how these three base classes are treated by or behave with respect to salient operators could be spelled out more explicitly.
(For example, I take it that an `Individual` should provide `.cross` and `.mutate`?

> In the codes, UserIndividual (respectively UserPopulation) is a container of elements in type of BinaryChromosome (respectively UserIndividual).

Could "the codes" be rephrased as "the example code"?
Additionally, it is unclear what the respectively parenthetical are trying to communicate.

---

Line 108: check the spacing between "Equation 1" and the ":" (ideally, there should not be a space here)

---

In the display math on line 118, could a semicolon be used to separate `x_i` and `i` rather than a comma (which is already separating 0/1 and 1/.../n)?

---

I take it that `element_class` is a pyrimidine API hook?

It may be easier on users less familiar with Python to use `#`'s to comment rather than `"""`'s.
Or put the `"""` content as a docstring on the class `UserIndividual`.

---

> ... `pop.solution` to decode the individual to the solution of the problem.

Is this meant to imply that there is a genotype (encoded)/phenotype (decoded) distinction within pyrimidine?
This hasn't yet been mentioned in the manuscript (or I missed it).

---

Given the discussion of pandas for the visualization/history API, a citation would be appropriate.
<https://pandas.pydata.org/about/citing.html>

---

Figure 1: It would be more standard (and easier to visualize) the standard deviation as an error band rather than plotting it on a separate axis. See, e.g, <https://stackoverflow.com/a/13157955/1733220>.

---

The first paragraph of the conclusion is a little vague and fluffy.
This could be made more specific, and could probably be just one sentence.
Also, the meaning of "experiments" is a bit unclear here.
Are you simply saying that the code has been used for lots of different projects?

---

Line 145. Use of "intelligent" here is also a bit strange. Perhaps "effective" or "powerful"?

---

> an interface that can be utilized at any time

Unclear what this means.

---

Like the first paragraph of the conclusion, the concluding paragraph could probably also be trimmed and tightened.
For instance, while I agree with the sentiment, the sentence "Certainly, there is ample room for further improvement." could probably be deleted entirely.
