### 2024-08-07 #6575 @boisgera
> I am unsure about the `formatting` of Pyrimidine in the title, as I don't see other JOSS papers with the title formatted this way.
> I didn't find any guidance in this regard in the JOSS Paper format guidelines.
> If this won't cause problems later, it's fine to keep it as is, but I just wanted to flag it so it could be fixed now if necessary (perhaps @boisgera could advise on this?).

Good catch! I couldn't tell if the use of the code font in the title could cause issues down the line. But at the very least, all recent papers accepted by JOSS mention the project name in the title *without* the backquotes. So I'd advocate to remove the backquotes in the paper and section titles (The track EiC would likely have the same request later in the process)

More generally, I wonder if the same change should not be generalized in the paper e.g. replace "`numpy`" with "NumPy".
My personal preference would be: use "NumPy" for a reference to the project and "`numpy`" for a reference to the Python object (a module) or the entry name in PyPi only.

I'll ask other editors what they think about this and report back here.
