### 2024-08-07 #6575 @boisgera
Update wrt formatting: definitely no backquote in the paper title.
