+++ paper/paper.bib
@@ -189,3 +189,31 @@ pages="560--564"
     title = {DEPP - Differential Evolution Parallel Program},
     journal = {Journal of Open Source Software}
 }
+
+@techreport{fogel,  
+  title = {Artificial intelligence through evolutionary programming},  
+  author = {Fogel, L.J. and Fogel, D.B.},  
+  institution = {U.S. Army Research Institute},  
+  year = {1986}, 
+  type = {Final Report}  
+}
+
+@article{kirkpatrick,
+  title={Optimization by Simulated Annealing},  
+  author={Kirkpatrick, S. and Gelatt, C. D., Jr. and Vecchi, M. P.},  
+  journal={Science},  
+  volume={220},  
+  number={4598},  
+  pages={671--679},  
+  year={1983}  
+}
+
+@InProceedings{mckinney,
+  author    = { {W}es {M}c{K}inney },
+  title     = { {D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython },
+  booktitle = { {P}roceedings of the 9th {P}ython in {S}cience {C}onference },
+  pages     = { 56 - 61 },
+  year      = { 2010 },
+  editor    = { {S}t\'efan van der {W}alt and {J}arrod {M}illman },
+  doi       = { 10.25080/Majora-92bf1922-00a }
+}
