+++ paper/paper.bib
@@ -217,3 +217,13 @@ pages="560--564"
   editor    = { {S}t\'efan van der {W}alt and {J}arrod {M}illman },
   doi       = { 10.25080/Majora-92bf1922-00a }
 }
+
+@article{kapitonova,
+title = {Algebraic programming: Methods and tools},
+author = {Kapitonova, Y.V. and Letichevskii, A.A.},
+journal = {Cybern Syst Anal},
+volume = {29},
+pages = {307--312},
+year = {1993},
+doi = {10.1007/BF01125535}
+}
