### 2024-09-09 #6575 @boisgera
Hi @Freakwill,

  - After a cursory look at the new version of your paper, I see that the new version of your paper was meant to address [Matthew's detailled comments](https://github.com/openjournals/joss-reviews/issues/6575#issuecomment-2270092590). (caveat: I have only sampled a few of the issues.)

 - As of today, there are no [open issues](https://github.com/Freakwill/pyrimidine/issues) in the project repository.

@mmore500, @sjvrijn can you consider the remaining items of your checklists ([1](https://github.com/openjournals/joss-reviews/issues/6575#issuecomment-2032735099)  and [2](https://github.com/openjournals/joss-reviews/issues/6575#issuecomment-2040630671))? Do you have any blocker?

Best regards,

Sébastien
