### 2024-09-22 04ee2ac
Clarify ambiguous "it"
