### 2024-09-22 fcdc889
Remove superfluous href
