### 2024-09-22 c3fb9fa
Clarify ambiguous sentence object
