### 2024-09-23 #6575 @mmore500
I am pleased to see the good progress on the manuscript, it has improved substantially. I opened a [pull request](https://github.com/Freakwill/pyrimidine/pull/6) with some suggestions on minor details in the manuscript. I also pointed out a few confusing sentences as a comment on that pull request [here](https://github.com/Freakwill/pyrimidine/pull/6#issuecomment-2367017288).

My major concerns at present are,
- It looks like the project's contiguous integration is currently broken. This is necessary to fix for the `Automated tests` checklist item.
- I have opened a few issues related to documentation.
- The manuscript is still too long, currently 1,600 words. (JOSS soft upper limit is 1,000 words.)

While having a very brief example of API usage can be useful in a JOSS paper, I believe that the second half of the paper would be more suited to a "quickstart" section in documentation. I would suggest that the author download and review a few recent articles from JOSS to get a sense for the content that typically is/isn't included in a JOSS manuscript.

Without the sections "An example to begin" and "Visualization," the manuscript would be closer to 1,250 words. Cutting these sections and a doing a little trimming/rewording elsewhere should be enough to get down closer to 1,000 words.

I have marked off a few more points on my checklist above. Ping me in this thread when I should return to continue my review!
