### 2024-09-23 #6575 @mmore500
Some of the references listed as "no DOI found" above are false negatives, e.g., <https://www.science.org/doi/10.1126/science.220.4598.671>. I would recommend going through these individually to see if a DOI is available.

crossref <https://github.com/Freakwill/pyrimidine/pull/6#issuecomment-2367017288>
