### 2024-09-30 #6575 @boisgera
AFAICT the project GitHub actions fail right now; I had a very quick look and apparently there was a `np.float_` error that was easy to fix but it was not the end of it (unknown `pd.chain` attribute?).

@Freakwill Could you ping us here when the actions pass again? Thanks!
