### 2024-11-25 #6575 @boisgera
I see that some PRs have been closed and that the CI works again, thanks for that @Freakwill!

With respect to the [latest comments of @mmore500](https://github.com/openjournals/joss-reviews/issues/6575#issuecomment-2367056960), AFAICT the remaining issues are:

-  Some issues relative to the docs are still open (I have not checked if the underlying issues have been solved). What is the status on this?

- The manuscript length has been reduced from ~1600 words to ~1400 words, but it's still significantly longer than the journal soft limit (1000 words, the range being 250-1000 words). Could you try to get it down to <1250 words please?  Since the focus of the paper should be ["an explanation of the software functionality and domain of use to a non-specialist reader"](https://joss.readthedocs.io/en/latest/paper.html#joss-paper-format) you may consider transferring part of the paper most technical content to either the project README or the project documentation.
