### 2025-01-17 #6575 @boisgera
Hi @Freakwill,

I see that you have shortened the paper (now 1155 words long) and resolved the remaining GitHub issues of your project. Thank you for that! 🙏

@sjvrijn, @mmore500 could you please reassess the state of the project and update your checklists accordingly?

Best regards,

Sébastien
