### 2025-02-05 #6575 @mmore500
I have completed my read-through of the paper and my checkboxes in the software paper section. I opened <https://github.com/Freakwill/pyrimidine/pull/12> with some final minor line editing suggestions. My final comments/suggestions on the manuscript are
- the use scenario/functionality described in the "Metaclasses" and "Mixin classes" sections are a bit challenging to decipher. I would revise these sections for concreteness/clarity, or (given that the manuscript is a bit long) replace these sections with an "Additional Features" section at the end where a one or two sentence description is given for each.
- the ending of the manuscript is very abrupt. Consider adding a brief concluding remark, even if it is just one sentence at the end.
