### 2025-02-05 #6575 @mmore500
✅  I have completed [my review checklist](https://github.com/openjournals/joss-reviews/issues/6575#issuecomment-2032735099).

I opened <https://github.com/Freakwill/pyrimidine/issues/13> with some remaining rendering and formatting issues in the documentation.

With respect to automated tests, it appears that CI is enabled via GitHub actions, but that the tests actually being skipped for most pushes to the repository--- see [here](https://github.com/Freakwill/pyrimidine/blob/940e8ec6171804ae880b135da9271d8f6736fa1d/.github/workflows/pytest.yml#L33) in the GitHub actions configuration yaml; I would strongly recommend deleting this if check that prevents the tests from being run and simply running the tests for all pushed commits (as they only take about 10 seconds to run anyways) and this would align better with standard practice.

I ran the unit tests locally and confirmed that they pass, so went ahead and checked the "Automated Tests" checkbox.

Given that all my boxes are checked, I believe that my portion of the review is now complete. Congratulations to the author, and thank you very much @boisgera for shepherding this process along. 👍
