### 2025-02-25 #6575 @boisgera
Hi @sjvrijn!

Given the changes in the paper and project, are you able to check all the items [in your checklist](https://github.com/openjournals/joss-reviews/issues/6575#issuecomment-2040630671)? Are there any remaining blockers that you'd like @Freakwill to address?

Best regards,

Sébastien
