### 2025-03-17 #6575 @boisgera
@sjvrijn, thanks a lot for your analysis and feedback! 🙏

@Freakwill, could you address the issues that Sander has opened?
