### 2025-04-07 #6575 @boisgera
> When my currently open PRs have been merged (or their underlying issues are otherwise addressed), that will concludes my review.

Great Sander, thanks a lot for this work! 🙏

If that's ok with you, I'll monitor the changes in the repo and check the last item "Functionality documentation" of your checklist when all the PRs/issues have been addressed.
