### 2025-05-12 #6575 @boisgera
Hi @Freakwill,

Coud you address the batch of issues/PRs provided by @sjvrijn so that we can move the review forrward?

Thanks in advance!
