### 2025-05-22 #6575 @boisgera
Hi @Freakwill,

Could you consider this:

> Hi [@Freakwill](https://github.com/Freakwill),
>
> Coud you address the batch of issues/PRs provided by [@sjvrijn](https://github.com/sjvrijn) so that we can move the review forward?
>
> Thanks in advance!

They are likely to be the last required steps required before we can move to the "post-review" stage.
