### 2025-06-12 #6575 @boisgera
Only two issues remaining:

  - https://github.com/Freakwill/pyrimidine/issues/24
  - https://github.com/Freakwill/pyrimidine/issues/21
