### 2025-08-04 #6575 @boisgera
@Freakwill

Please consider this (formatting) PR:

  - https://github.com/Freakwill/pyrimidine/pull/26
