### 2025-08-04 #6575 @boisgera
Afterwards, could you have a look at [Additional Author Tasks After Review is Complete](https://github.com/openjournals/joss-reviews/issues/6575#issuecomment-3149322329) please?

> [@Freakwill](https://github.com/Freakwill)
>
> Please consider this (formatting) PR:
>
>     * [Update spacing in references Freakwill/pyrimidine#26](https://github.com/Freakwill/pyrimidine/pull/26)
