### 2025-08-26 #6575 @boisgera
Hi @Freakwill!

Could you have a look at [Additional Author Tasks After Review is Complete](https://github.com/openjournals/joss-reviews/issues/6575#issuecomment-3149322329) please?

  - **Archive the release on Zenodo/figshare/etc and post the DOI here.**

I'll need the DOI of your archive in order to proceed with the review.
