### 2025-09-04 #6575 @boisgera
Hi @Freakwill,

Some minor issue with the Zenodo archive:

  - Could you please list only yourself as the author? (Reviewers and editors will be acknowledged differently in the final published pdf)

  - Could you declare your ORCID so that you get the ORCID green badge/link added to your name ? (see for example [here](https://zenodo.org/records/14723141))

  - Please use exactly the title of your paper ("Pyrimidine: An algebra-inspired Programming framework for evolutionary algorithms") as the Zenodo title (no author, no version there, etc.)

  - Please sync the Zenodo version tag with your software version (here "0.18" instead of "first" ; that's what you see in the right column on your Zenodo page)
